Cog creation and chat extension
Tested first cog in chat.py
Moved general commands from bot.py to chat.py
